+++ Changelog.md
@@ -1,3 +1,6 @@
 # Changelog
 
 ## Version 0.1.0
+* added variational quantum classifier
+* added encodings
+* added training